Raise an error if the Datapoint API returns a 400 or 500 status
If the Datapoint API returns a status code of 4xx or 5xx, an error
should be raised straightaway. By raising the error as soon as it
occurs, the user can see the HTTP error message, and so work out why the
request failed. If an error is not raised at this point, the program
will continue to execute and will probably give a JSON parsing error
much further down the line, which is both inefficient and harder to
debug.