Handle request for obs when site ID is invalid
If a request is made to get all observations and the site ID is not
recognised, no error is raised by the Datapoint API. Instead, the JSON
that is returned simply doesn't contain any observation data. This is
now handled by raising a newly created type of error, a SiteError, with
a message.