Add migration for Site table
The primary key is the id column, which is the id of the site as given
by DataPoint. This has a not null constraint (because it is the primary
key), but the server default of null is added in order to prevent a
default integer being used. Having a not null constraint together with
a server default of null means that the site id must be explicitly given
by the user.