Show site name on page displaying obs for a site
The site name is fetched from the database to be displayed on this page.
The alternative way of getting the site name would be to change what the
DatapointClient returns, but since I want a single source of truth for
the sites, I am getting the name from the database and accepting the
slight performance hit that this will bring.