Add boolean 'observations' column to site table
Every site in the Site table has forecast data, but the purpose of the
new column is to show whether a site also has observation data - only a
small number of the sites have observation data as well as forecast
data.

This column is indexed as it will be common to search for sites where
observations is True. Using a partial index on the column was tried, but
was actually slowly than no index.