Lookup observation sites from the database
The list of observation sites was previously hard coded in a file. This has
been replaced with a dao function which gets the sites with observations
from the database instead. This isn't fully tested yet (ok, let's not
lie, it's not tested at all) because I need to think through how to test
functions which hit the database given that the database is essentially
a read-only database with data that never changes.