Use application factory pattern
Changed to use the application factory pattern in
order to be able to add different config for different
environments later. As part of this change, added in a
single blueprint, 'main', so that I don't need to include
every view function in the create_app function I can just
register the new blueprint.