Set up different test and dev config
Separate test and development config was needed because all tests
were having to run against the main database, which is dependent
on the data import scripts having been run. As well as setting up
separate config classes, added fixtures to allow a test database to
be used and cleaned at the end of every test.