Add a page to display the details for a site
This displays the latitude, longitude and elevation. It links to the
forecast for that site and the observations too (if available).

This page is now linked to from the search box on the homepage and the
page listing all observations, and the search box has been added to the
page listing all forecasts too.