Customise links on results page
The search bar is used in three places and always uses 'results.html' to
display the results. Each result on this page is a link, but the
location of the link should vary according to which page you've just come
from. (Previously, each result on the results page always linked to the
'site_observation' page.)

This adds a hidden field to the search form - the result of this
determines where each search result on the page should link to and where
the back button on the page should link to.