Add maps to site page
Changed the site page to include a couple of maps - a large detailed
map, and a smaller overview map. This uses Leaflet for the maps and the
ESRI Leaflet plugin for the basemaps to avoid having to use a tile
server that requires an API key.

This commit involves JavaScript ☠️